fix: changed folder structure
