--- BackEnd/src/controllers/authController.ts
@@ -1,51 +1,0 @@
-import { Request, Response } from 'express';
-import AuthService from '../services/authService';
-
-class AuthController {
-    private authService: AuthService;
-
-    constructor() {
-        this.authService = new AuthService();
-    }
-
-    public registerUser = async (req: Request, res: Response) => {
-        try {
-            const user = await this.authService.register(req.body);
-            res.status(201).json(user);
-        } catch (error) {
-            if (error instanceof Error) {
-                res.status(400).json({ message: error.message });
-            } else {
-                res.status(400).json({ message: 'An unknown error occurred' });
-            }
-        }
-    };
-
-    public loginUser = async (req: Request, res: Response) => {
-        try {
-            const token = await this.authService.login(req.body);
-            res.status(200).json({ token });
-        } catch (error) {
-            if (error instanceof Error) {
-                res.status(400).json({ message: error.message });
-            } else {
-                res.status(400).json({ message: 'An unknown error occurred' });
-            }
-        }
-    };
-
-    public googleAuth = async (req: Request, res: Response) => {
-        try {
-            const token = await this.authService.googleAuth(req.body);
-            res.status(200).json({ token });
-        } catch (error) {
-            if (error instanceof Error) {
-                res.status(400).json({ message: error.message });
-            } else {
-                res.status(400).json({ message: 'An unknown error occurred' });
-            }
-        }
-    };
-}
-
-export default AuthController;
