--- BackEnd/src/routes/authRoutes.ts
@@ -1,72 +1,0 @@
-import { Router } from 'express';
-import AuthController from '../controllers/authController';
-
-const router = Router();
-const authController = new AuthController();
-
-
-/**
- * @swagger
- * /api/auth/register:
- *   post:
- *     summary: Register a new user
- *     tags: [Auth]
- *     requestBody:
- *       required: true
- *       content:
- *         application/json:
- *           schema:
- *             type: object
- *             properties:
- *               email:
- *                 type: string
- *               password:
- *                 type: string
- *     responses:
- *       201:
- *         description: User registered successfully
- *       400:
- *         description: Bad request
- */
-router.post('/register', authController.registerUser);
-
-/**
- * @swagger
- * /api/auth/login:
- *   post:
- *     summary: Login a user
- *     tags: [Auth]
- *     requestBody:
- *       required: true
- *       content:
- *         application/json:
- *           schema:
- *             type: object
- *             properties:
- *               email:
- *                 type: string
- *               password:
- *                 type: string
- *     responses:
- *       200:
- *         description: User logged in successfully
- *       400:
- *         description: Bad request
- */
-router.post('/login', authController.loginUser);
-
-/**
- * @swagger
- * /api/auth/google:
- *   get:
- *     summary: Authenticate with Google
- *     tags: [Auth]
- *     responses:
- *       200:
- *         description: User authenticated successfully
- *       400:
- *         description: Bad request
- */
-router.get('/google', authController.googleAuth);
-
-export default router;
