--- BackEnd/src/routes/userRoutes.ts
@@ -1,53 +1,0 @@
-import { Router } from 'express';
-import UserController from '../controllers/userController';
-
-const router = Router();
-const userController = new UserController();
-
-/**
- * @swagger
- * /api/user/profile:
- *   get:
- *     summary: Get user profile
- *     tags: [User]
- *     parameters:
- *       - in: query
- *         name: id
- *         schema:
- *           type: string
- *         required: true
- *         description: The user ID
- *     responses:
- *       200:
- *         description: User profile retrieved successfully
- *       400:
- *         description: Bad request
- */
-router.get('/profile', userController.getUserProfile);
-
-/**
- * @swagger
- * /api/user/profile:
- *   put:
- *     summary: Update user profile
- *     tags: [User]
- *     requestBody:
- *       required: true
- *       content:
- *         application/json:
- *           schema:
- *             type: object
- *             properties:
- *               email:
- *                 type: string
- *               password:
- *                 type: string
- *     responses:
- *       200:
- *         description: User profile updated successfully
- *       400:
- *         description: Bad request
- */
-router.put('/profile', userController.updateUserProfile);
-
-export default router;
